Version 4.5

+++ events/guildDelete.js
@@ -6,7 +6,10 @@ module.exports = class {
   }
 
   async run(guild) {
-
+    if (!guild.available) return;
+    const def = await this.client.settings.get("default").getField("settings").run();
+    this.client.user.setPresence({game: {name: `${def.prefix}help | ${this.client.guilds.size} Servers`, type:0}});
+    
     // Well they're gone. Let's remove them from the settings!
     this.client.settings.get(guild.id).delete().run();
   }
